requirements.txt
web3==6.19.0
eth-account==0.13.3
python-dotenv==1.0.1
